+++ app/category/routers_category.py
@@ -4,7 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from category.model_category import CategoryModel
 from routers.consts import RouteSlug
 from models.database import get_session
-from category.category_schema import CategoryOptionalSchema
+from category.schema_category import CategoryOptionalSchema
 from services.service_service import CategoryService
 from dependencies.base_dependency import (
     ValidGetByIdDependency,
