+++ app/includer/routers.py
@@ -1,4 +1,4 @@
-from category.category_routers import router_category
+from category.routers_category import router_category
 
 routers_all2 = [
     router_category,
